feat(button-basic): change basic button styles to theme friendly
