+++ src/assets/effects/theme.scss
@@ -1,0 +1,13 @@
+@mixin themeify($themes) {
+  @each $theme-name, $theme-map in $themes {
+    $theme-map: $theme-map !global;
+
+    body[data-theme="#{$theme-name}"] {
+      @content;
+    }
+  }
+}
+
+@function t($key) {
+  @return map-get($theme-map, $key);
+}
