+++ src/components/ButtonBasic/dim.scss
@@ -1,0 +1,8 @@
+$btn-corner-radius: 0.8rem;
+$btn-line-height: 1;
+$btn-font-size: 1.6rem;
+$btn-padding: (
+  small: 0.8rem,
+  medium: 1rem,
+  large: 1.2rem,
+);
