+++ src/views/pageTypewriter/components/Portrait/TransformSwitcher/TransformSwitcher.vue
@@ -4,7 +4,7 @@ import { ConfigConvertDirectionEnum } from "@/shared";
 import { useConfig, useFontSelector } from "../../../hooks";
 import variables from "./variables.module.scss";
 import SvgIcon from "@/components/SvgIcon.vue";
-import ButtonBasic from "@/components/ButtonBasic.vue";
+import ButtonBasic from "@/components/ButtonBasic/ButtonBasic.vue";
 import ImgContentArrow from "../../../assets/convert-arrow.svg";
 
 const { config } = useConfig();
